fixup! Add created_at to WorkItem API (#506)
