+++ design/media_types.go
@@ -37,6 +37,7 @@ var AuthToken = a.MediaType("application/vnd.authtoken+json", func() {
 })
 
 // workItem is the media type for work items
+// Deprecated, but kept around as internal model for now.
 var workItem = a.MediaType("application/vnd.workitem+json", func() {
 	a.TypeName("WorkItem")
 	a.Description("A work item hold field values according to a given field type")
@@ -69,32 +70,6 @@ var meta = a.Type("workItemListResponseMeta", func() {
 	a.Attribute("totalCount", d.Integer)
 
 	a.Required("totalCount")
-})
-
-// workItemListResponse contains paged results for listing work items and paging links
-var workItemListResponse = a.MediaType("application/vnd.workitemlist+json", func() {
-	a.TypeName("WorkItemListResponse")
-	a.Description("Holds the paginated response to a work item list request")
-	a.Attribute("links", pagingLinks)
-	a.Attribute("meta", meta)
-	a.Attribute("data", a.CollectionOf(workItem))
-
-	a.Required("links")
-	a.Required("meta")
-	a.Required("data")
-
-	a.View("default", func() {
-		a.Attribute("links", func() {
-			a.Attribute("prev", d.String)
-			a.Attribute("next", d.String)
-			a.Attribute("first", d.String)
-			a.Attribute("last", d.String)
-		})
-		a.Attribute("meta", func() {
-			a.Attribute("totalCount", d.Number)
-		})
-		a.Attribute("data")
-	})
 })
 
 // fieldDefinition defines the possible values for a field in a work item type
