+++ design/resources.go
@@ -5,91 +5,6 @@ import (
 	a "github.com/goadesign/goa/design/apidsl"
 )
 
-var _ = a.Resource("workitem", func() {
-	a.BasePath("/workitems")
-
-	a.Action("show", func() {
-		a.Routing(
-			a.GET("/:id"),
-		)
-		a.Description("Retrieve work item with given id.")
-		a.Params(func() {
-			a.Param("id", d.String, "id")
-		})
-		a.Response(d.OK, func() {
-			a.Media(workItem)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.NotFound, JSONAPIErrors)
-	})
-
-	a.Action("list", func() {
-		a.Routing(
-			a.GET(""),
-		)
-		a.Description("List work items.")
-		a.Params(func() {
-			a.Param("filter", d.String, "a query language expression restricting the set of found work items")
-			a.Param("page", d.String, "Paging in the format <start>,<limit>")
-		})
-		a.Response(d.OK, func() {
-			a.Media(a.CollectionOf(workItem))
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-	})
-
-	a.Action("create", func() {
-		a.Security("jwt")
-		a.Routing(
-			a.POST(""),
-		)
-		a.Description("create work item with type and id.")
-		a.Payload(CreateWorkItemPayload)
-		a.Response(d.Created, "/workitems/.*", func() {
-			a.Media(workItem)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.Unauthorized, JSONAPIErrors)
-	})
-	a.Action("delete", func() {
-		a.Security("jwt")
-		a.Routing(
-			a.DELETE("/:id"),
-		)
-		a.Description("Delete work item with given id.")
-		a.Params(func() {
-			a.Param("id", d.String, "id")
-		})
-		a.Response(d.OK)
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.Unauthorized, JSONAPIErrors)
-	})
-	a.Action("update", func() {
-		a.Security("jwt")
-		a.Routing(
-			a.PUT("/:id"),
-		)
-		a.Description("update the given work item with given id.")
-		a.Params(func() {
-			a.Param("id", d.String, "id")
-		})
-		a.Payload(UpdateWorkItemPayload)
-		a.Response(d.OK, func() {
-			a.Media(workItem)
-		})
-		a.Response(d.BadRequest, JSONAPIErrors)
-		a.Response(d.InternalServerError, JSONAPIErrors)
-		a.Response(d.NotFound, JSONAPIErrors)
-		a.Response(d.Unauthorized, JSONAPIErrors)
-	})
-
-})
-
 var _ = a.Resource("workitemtype", func() {
 
 	a.BasePath("/workitemtypes")
