+++ main.go
@@ -182,9 +182,6 @@ func main() {
 	workitemCtrl := NewWorkitemController(service, appDB)
 	app.MountWorkitemController(service, workitemCtrl)
 
-	workitem2Ctrl := NewWorkitem2Controller(service, appDB)
-	app.MountWorkitem2Controller(service, workitem2Ctrl)
-
 	// Mount "workitemtype" controller
 	workitemtypeCtrl := NewWorkitemtypeController(service, appDB)
 	app.MountWorkitemtypeController(service, workitemtypeCtrl)
